// src/model/basicRemasterResponse.ts
/**
 * Generated by orval v6.22.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { BasicUserResponse } from "./basicUserResponse";

export interface BasicRemasterResponse {
  createdAt: string;
  id: string;
  name: string;
  url: string;
  user: BasicUserResponse;
}